move graceful shutdown helper to its own interface
